requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
websockets==12.0
google-generativeai==0.8.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
google-cloud-secret-manager==2.20.0
google-cloud-tasks==2.16.0
vertexai==1.70.0
python-dotenv==1.0.0
httpx==0.27.0
